+++ src/main/java/com/brainsmash/broken_world/Client.java
@@ -1,5 +1,6 @@
 package com.brainsmash.broken_world;
 
+import com.brainsmash.broken_world.blocks.multiblock.MultiblockUtil;
 import com.brainsmash.broken_world.entity.impl.EntityDataExtension;
 import com.brainsmash.broken_world.entity.impl.PlayerDataExtension;
 import com.brainsmash.broken_world.items.weapons.guns.GunItem;
@@ -76,7 +77,7 @@ public class Client implements ClientModInitializer {
         EntityRegister.registEntitiesClientSide();
         ItemRegister.registItemClientSide();
 
-        BlockRegister.registMultiblockClientSide();
+        MultiblockUtil.registMultiblockClientSide();
 
         FluidRegister.RegistFluidClientSide();
